+++ .github/scripts/generate_reference.py
@@ -50,7 +50,7 @@ thermo:
 
 dft:
   func_basis: wb97m-v/def2-tzvpd
-  max_cycle: null
+  max_cycle: 100
   conv_tol: 1.0e-9
   grid_level: 3
 """